chore: back: 여행지 추천 디버그

+++ back/src/main/java/com/ssafy/daero/common/util/DistanceUtil.java
@@ -5,7 +5,7 @@ public class DistanceUtil {
         private ReverseHaversine() {
         }
 
-        public static double[] calculateCoordinate(double latitude, double longitude, int distance) {
+        public static double[] calculateCoordinate(double latitude, double longitude, double distance) {
             double dif = distance * 0.007;
             return new double[]{latitude - dif, latitude + dif, longitude - dif, longitude + dif};
         }
